adição da biblioteca de classes
pasta com domínios entidades e classes auxiliares
