+++ biblioteca_de_classes/entidades.cpp
@@ -1,0 +1,1 @@
+#include "entidades.hpp"
