--- entidades.cpp
@@ -1,1 +1,0 @@
-#include "entidades.hpp"
